refactor: replace BigDecimal with Double for financial calculations in DTOs and use cases

+++ src/main/java/com/jf/PetApp/application/investment/dto/AllocationSliceDTO.java
@@ -2,11 +2,9 @@ package com.jf.PetApp.application.investment.dto;
 
 import com.jf.PetApp.core.domain.enums.InvestmentType;
 
-import java.math.BigDecimal;
-
 public record AllocationSliceDTO(
     InvestmentType type,
-    BigDecimal currentValue,
-    BigDecimal portfolioPercent
+    Double currentValue,
+    Double portfolioPercent
 ) {
 }
