+++ src/main/java/com/jf/PetApp/application/investment/dto/PortfolioHistoryPointDTO.java
@@ -1,11 +1,10 @@
 package com.jf.PetApp.application.investment.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record PortfolioHistoryPointDTO(
     LocalDate date,
-    BigDecimal investedCapital,
-    BigDecimal portfolioValue
+    Double investedCapital,
+    Double portfolioValue
 ) {
 }
